blacklistfilter: watcher: Adapt for IPv6, Pure c++ compilation
