--- blacklistfilter/blacklist_watcher.c
@@ -1,123 +1,0 @@
-/* This code is highly inspired by this source:
-   http://man7.org/linux/man-pages/man7/inotify.7.html */
-
-#include <sys/inotify.h>
-#include <stdio.h>
-#include <errno.h>
-#include <unistd.h>
-#include <pthread.h>
-#include <poll.h>
-
-#include "blacklist_watcher.h"
-
-#ifdef DEBUG
-#define DBG(x) fprintf x;
-#else
-#define DBG(x)
-#endif
-
-extern int stop;
-extern int BL_RELOAD_FLAG;
-extern pthread_mutex_t BLD_SYNC_MUTEX;
-
-/**
- * \brief Handles inotify events occuring on the filedescriptor.
- * \param fd File descriptor to watch for events
- */
-static void handle_events(int fd)
-{
-    /* Some systems cannot read integer variables if they are not
-       properly aligned. On other systems, incorrect alignment may
-       decrease performance. Hence, the buffer used for reading from
-       the inotify file descriptor should have the same alignment as
-       struct inotify_event. */
-    char buf[4096] __attribute__ ((aligned(__alignof__(struct inotify_event))));
-
-    const struct inotify_event *event;
-    ssize_t len;
-    char *ptr;
-
-    /* Loop while events can be read from inotify file descriptor. */
-    while (1) {
-        len = read(fd, buf, sizeof(buf));
-        if (len == -1 && errno != EAGAIN) {
-            perror("Error: Couldnt read from fd");
-            stop = 1; return;
-        }
-
-        /* If the nonblocking read() found no events to read, then
-           it returns -1 with errno set to EAGAIN. In that case,
-           we exit the loop. */
-        if (len <= 0) {
-            break;
-        }
-
-        /* Loop over all events in the buffer */
-        for (ptr = buf; ptr < buf + len; ptr += sizeof(struct inotify_event) + event->len) {
-            event = (const struct inotify_event *) ptr;
-
-            if (event->mask & IN_CLOSE_WRITE) {
-                DBG((stderr, "Blacklist watcher setting a flag to reload blacklists\n"));
-                pthread_mutex_lock(&BLD_SYNC_MUTEX);
-                BL_RELOAD_FLAG = 1;
-                pthread_mutex_unlock(&BLD_SYNC_MUTEX);
-            }
-        }
-    }
-}
-
-
-/**
- * \brief Watch blacklist files for IN_CLOSE_WRITE event
- * and set appropriate flag if this file changes
- * \param arg File name of the blacklist to watch
- */
-void *watch_blacklist_files(void *arg)
-{
-    const char *bl_file = (const char *) arg;
-    int fd, poll_num;
-    int wd; // TODO: is just one watch descriptor fine?
-    nfds_t nfds;
-    struct pollfd fds[1];
-
-    /* Create the file descriptor for accessing the inotify API */
-    fd = inotify_init1(IN_NONBLOCK);
-    if (fd == -1) {
-        perror("Error: Couldnt initialize inotify");
-        stop = 1; return NULL;
-    }
-
-    /* Watch the file for IN_CLOSE_WRITE event */
-    wd = inotify_add_watch(fd, bl_file, IN_CLOSE_WRITE);
-    if (wd == -1) {
-        perror("Error: Cannot watch the file, inotify_add_watch failed");
-        stop = 1; return NULL;
-    }
-
-    /* Prepare for polling */
-    nfds = 1;
-
-    fds[0].fd = fd;
-    fds[0].events = POLLIN;
-
-    DBG((stderr, "Blacklist watcher listening for changes in file %s\n", bl_file));
-
-    while (1) {
-        poll_num = poll(fds, nfds, -1);
-        if (poll_num == -1) {
-            if (errno == EINTR) {
-                continue;
-            }
-
-            perror("Error: Poll failure");
-            stop = 1; return NULL;
-        }
-
-        if (poll_num > 0) {
-            if (fds[0].revents & POLLIN) {
-                /* Inotify events are available */
-                handle_events(fd);
-            }
-        }
-    }
-}
