+++ blacklistfilter/blacklist_watcher.cpp
@@ -1,24 +1,14 @@
+/* This code is highly inspired by this source:
+   http://man7.org/linux/man-pages/man7/inotify.7.html */
 
 #include <sys/inotify.h>
-#include <string>
-#include <cctype>
-#include <algorithm>
-#include <iostream>
-#include <fstream>
-#include <cstdlib>
-#include <map>
-#include <vector>
-#include <stdint.h>
-#include <signal.h>
-#include <getopt.h>
-#include <dirent.h>
+#include <stdio.h>
+#include <errno.h>
 #include <unistd.h>
 #include <pthread.h>
-#include <nemea-common/nemea-common.h>
-#include <unirec/unirec.h>
-#include <libtrap/trap.h>
 #include <poll.h>
 
+#include "ipblacklistfilter.h"
 #include "blacklist_watcher.h"
 
 #ifdef DEBUG
@@ -27,11 +17,13 @@
 #define DBG(x)
 #endif
 
-using namespace std;
-
+extern int stop;
 extern int BL_RELOAD_FLAG;
-extern pthread_mutex_t BLD_SYNC_MUTEX;
 
+/**
+ * \brief Handles inotify events occuring on the filedescriptor.
+ * \param fd File descriptor to watch for events
+ */
 static void handle_events(int fd)
 {
     /* Some systems cannot read integer variables if they are not
@@ -39,37 +31,29 @@ static void handle_events(int fd)
        decrease performance. Hence, the buffer used for reading from
        the inotify file descriptor should have the same alignment as
        struct inotify_event. */
-
-    char buf[4096]
-            __attribute__ ((aligned(__alignof__(struct inotify_event))));
+    char buf[4096] __attribute__ ((aligned(__alignof__(struct inotify_event))));
 
     const struct inotify_event *event;
     ssize_t len;
     char *ptr;
 
     /* Loop while events can be read from inotify file descriptor. */
-
-    for (;;) {
-        /* Read some events. */
-
+    while (1) {
         len = read(fd, buf, sizeof(buf));
         if (len == -1 && errno != EAGAIN) {
-            perror("read");
-            exit(EXIT_FAILURE);
+            perror("Error: Couldnt read from fd");
+            stop = 1; return;
         }
 
         /* If the nonblocking read() found no events to read, then
            it returns -1 with errno set to EAGAIN. In that case,
            we exit the loop. */
-
-        if (len <= 0)
+        if (len <= 0) {
             break;
+        }
 
         /* Loop over all events in the buffer */
-
-        for (ptr = buf; ptr < buf + len;
-             ptr += sizeof(struct inotify_event) + event->len) {
-
+        for (ptr = buf; ptr < buf + len; ptr += sizeof(struct inotify_event) + event->len) {
             event = (const struct inotify_event *) ptr;
 
             if (event->mask & IN_CLOSE_WRITE) {
@@ -83,55 +67,52 @@ static void handle_events(int fd)
 }
 
 
-/* Watch directory with blacklist files for IN_CLOSE_WRITE event
- * and set appropriate flag if these files change */
-void * watch_blacklist_files(void * arg)
+/**
+ * \brief Watch blacklist files for IN_CLOSE_WRITE event
+ * and set appropriate flag if this file changes
+ * \param arg File name of the blacklist to watch
+ */
+void *watch_blacklist_files(void *arg)
 {
-    const char * bl_file = reinterpret_cast<string*>(arg)->c_str();
+    const char *ipv4_file = ((config_t *) arg) -> ipv4_blacklist_file;
+    const char *ipv6_file = ((config_t *) arg) -> ipv6_blacklist_file;
     int fd, poll_num;
-    int wd; // TODO: is just one watch descriptor fine?
+    int wd_4, wd_6;
     nfds_t nfds;
     struct pollfd fds[1];
 
     /* Create the file descriptor for accessing the inotify API */
-
     fd = inotify_init1(IN_NONBLOCK);
     if (fd == -1) {
-        perror("inotify_init1");
-        exit(EXIT_FAILURE);
+        perror("Error: Couldnt initialize inotify");
+        stop = 1; return NULL;
     }
 
-    /* Mark directories for events
-       - file was opened
-       - file was closed */
-
-    wd = inotify_add_watch(fd, bl_file, IN_CLOSE_WRITE);
-    if (wd == -1) {
-        fprintf(stderr, "Cannot watch '%s'\n", bl_file);
-        perror("inotify_add_watch");
-        exit(EXIT_FAILURE);
+    /* Watch the files for IN_CLOSE_WRITE event */
+    wd_4 = inotify_add_watch(fd, ipv4_file, IN_CLOSE_WRITE);
+    wd_6 = inotify_add_watch(fd, ipv6_file, IN_CLOSE_WRITE);
+    if (wd_4 == -1 || wd_6 == -1) {
+        perror("Error: Cannot watch the file, inotify_add_watch failed");
+        stop = 1; return NULL;
     }
 
     /* Prepare for polling */
-
     nfds = 1;
-
-    /* Console input */
 
     fds[0].fd = fd;
     fds[0].events = POLLIN;
 
-    /* Wait for events and/or terminal input */
-
-    DBG((stderr, "Blacklist watcher listening for changes in file %s\n", bl_file));
+    DBG((stderr, "Blacklist watcher listening for changes in files %s, %s\n", ipv4_file, ipv6_file));
 
     while (1) {
         poll_num = poll(fds, nfds, -1);
         if (poll_num == -1) {
-            if (errno == EINTR)
+            if (errno == EINTR) {
                 continue;
-            perror("poll");
-            exit(EXIT_FAILURE);
+            }
+
+            perror("Error: Poll failure");
+            stop = 1; return NULL;
         }
 
         if (poll_num > 0) {
