+++ blacklistfilter/blacklist_watcher.h
@@ -1,14 +1,11 @@
 #ifndef BLACKLISTFILTER_BLACKLIST_WATCHER_H
 #define BLACKLISTFILTER_BLACKLIST_WATCHER_H
 
-#ifdef __cplusplus
-extern "C" {
-#endif
+/**
+* Mutex for synchronization.
+*/
+static pthread_mutex_t BLD_SYNC_MUTEX;
 
 void *watch_blacklist_files(void *);
-
-#ifdef __cplusplus
-}
-#endif
 
 #endif //BLACKLISTFILTER_BLACKLIST_WATCHER_H
