--- blacklistfilter/blacklist_watcher/blacklist_watcher.h
@@ -1,10 +1,0 @@
-//
-// Created by filip on 10.7.18.
-//
-
-#ifndef BLACKLISTFILTER_BLACKLIST_WATCHER_H
-#define BLACKLISTFILTER_BLACKLIST_WATCHER_H
-
-void * watch_blacklist_files(void *);
-
-#endif //BLACKLISTFILTER_BLACKLIST_WATCHER_H
